telegram_downloader/requirements.txt
PySide6>=6.5.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pyautogui>=0.9.54
mss>=9.0.0
pynput>=1.7.6